Se separa el test en una clase individual
